Demo/prototype for finding distance from angles
This code demonstrates the method of using two angles (potentially read from two optical shaft encoders) to find the distance of an object that they are both pointing at. The code I made that turns this in to a class got corrupted, so this is all I have for that :(, but this works as a proof of concept that we can turn into other languages.